% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_antigen_list}
\alias{build_antigen_list}
\title{Build Antigen List}
\usage{
build_antigen_list(antigen_to_eval = c("ALL"), verbose = TRUE)
}
\arguments{
\item{antigen_to_eval}{Character vector of requested antigens (or "ALL").
Defaults to "ALL".}

\item{verbose}{Logical; whether to message progress. Default = TRUE.}
}
\value{
Character vector of antigens to evaluate. If none, returns character(0).
}
\description{
Utility function to derive the set of antigens to evaluate, given a target list
and the system defaults. Handles "ALL" and live-virus logic.
}
\keyword{internal}
