% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mon_with_grace}
\alias{mon_with_grace}
\title{Convert months to days with optional grace period}
\usage{
mon_with_grace(x, grace = 4)
}
\arguments{
\item{x}{Number of months}

\item{grace}{Days to subtract (default 4)}
}
\value{
Days
}
\description{
Convert months to days with optional grace period
}
\keyword{internal}
