% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{month_from_date}
\alias{month_from_date}
\title{Extract Calendar Month from Date}
\usage{
month_from_date(d)
}
\arguments{
\item{d}{A vector of class \code{Date}.}
}
\value{
An integer vector of calendar months (1-12).
}
\description{
Internal helper to return the calendar month (1-12) as an integer from a
\code{Date} vector.
}
\examples{
VaccineValidateR:::month_from_date(as.Date(c("2000-01-01", "2025-09-18")))

}
\keyword{internal}
