% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{next_sept1}
\alias{next_sept1}
\title{Compute Next September 1 After a Given Date}
\usage{
next_sept1(d)
}
\arguments{
\item{d}{A vector of class \code{Date}.}
}
\value{
A vector of \code{Date} values, each corresponding to
the September 1 following (or equal to) the year of \code{d}.
}
\description{
Given a vector of \code{Date} objects, returns the next September 1
on or after each input date. Used internally for immunization schedule
calculations (e.g., seasonal influenza logic).
}
\examples{
# Internal use only
test_dates <- as.Date(c("2020-01-15", "2020-09-01", "2020-12-15"))
VaccineValidateR:::next_sept1(test_dates)

}
\keyword{internal}
