% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{RSV_DATE}
\alias{RSV_DATE}
\title{Vaccine schedule reference dates}
\format{
Constants of class \code{Date}.
\describe{
\item{RSV_DATE}{2023-09-01, Beyfortus release (prior = palivizumab).}
\item{MENB_DATE}{2024-10-24, MenB guidance change.}
\item{OPV_DATE}{2016-04-01, OPV rule update.}
\item{HPV_DATE}{2016-04-01, HPV schedule update.}
\item{PCV_DATE}{2010-02-24, PCV13 release.}
}
}
\usage{
RSV_DATE
}
\description{
Important implementation or guidance change dates used in dose validation.
}
\keyword{internal}
