% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SEASONAL_ANTIGENS}
\alias{SEASONAL_ANTIGENS}
\title{Seasonal Antigens}
\format{
A character vector of antigen names.
}
\usage{
SEASONAL_ANTIGENS
}
\description{
Internal constant listing antigens that follow seasonal rules
(e.g., influenza, COVID, RSV).
}
\examples{
VaccineValidateR:::SEASONAL_ANTIGENS

}
\keyword{internal}
