% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{supported_antigens}
\alias{supported_antigens}
\title{Returns the list of Antigens the system is capable of interpreting}
\usage{
supported_antigens()
}
\value{
Character vector of the system antigens
}
\description{
Returns the list of Antigens the system is capable of interpreting
}
