% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SYSTEM_ANTIGENS}
\alias{SYSTEM_ANTIGENS}
\title{Supported antigens}
\format{
An object of class \code{character} of length 15.
}
\usage{
SYSTEM_ANTIGENS
}
\value{
Character vector of antigen names.
}
\description{
List of antigens the package is capable of interpreting. Used internally.
}
\keyword{internal}
