% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{VISIT_RETURN_COLUMNS}
\alias{VISIT_RETURN_COLUMNS}
\title{Visit-level return columns}
\format{
A character vector of column names.
}
\usage{
VISIT_RETURN_COLUMNS
}
\description{
Names of the standard columns returned by \code{evaluate_visits()}.
}
\keyword{internal}
