% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{VISIT_RETURN_DEFS}
\alias{VISIT_RETURN_DEFS}
\title{Visit-level column definitions}
\format{
A named character vector where names correspond to column names
  and values are descriptions.
}
\usage{
VISIT_RETURN_DEFS
}
\description{
Human-readable definitions of each column in \code{VISIT_RETURN_COLUMNS}.
}
\keyword{internal}
