% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wk_with_grace}
\alias{wk_with_grace}
\title{Convert weeks to days with optional grace period}
\usage{
wk_with_grace(x, grace = 4)
}
\arguments{
\item{x}{Number of weeks}

\item{grace}{Days to subtract (default 4)}
}
\value{
Days
}
\description{
Convert weeks to days with optional grace period
}
\keyword{internal}
