% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{year_from_date}
\alias{year_from_date}
\title{Extract Calendar Year from Date}
\usage{
year_from_date(d)
}
\arguments{
\item{d}{A vector of class \code{Date}.}
}
\value{
An integer vector of calendar years.
}
\description{
Internal helper to return the calendar year as an integer from a
\code{Date} vector.
}
\examples{
VaccineValidateR:::year_from_date(as.Date(c("2000-01-01", "2025-09-18")))

}
\keyword{internal}
