% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yr_with_grace}
\alias{yr_with_grace}
\title{Convert years to days with optional grace period}
\usage{
yr_with_grace(x, grace = 4)
}
\arguments{
\item{x}{Number of years}

\item{grace}{Days to subtract (default 4)}
}
\value{
Days
}
\description{
Convert years to days with optional grace period
}
\keyword{internal}
